drawer changed to permanent mini variant
